Can change min_colour_clustering_points
When analysing plates with two colours of cells, the automatic
colour recognition box may be selected. Clustering may be modified
in two ways, by changing the minimum search distance between points
and by changing how many points need to be together to be considered
a cluster.

Up to now, the minimum number of points was fixed. Now it can be
varied between 4 and 10 with a spin button. Higher numbers mean
less of a chance of joining two separate clusters.

Also included in this commit is the commented code to sort colour
clusters by the abundance of colonies in the cluster. This code
will be uncommented in the next commit