Bugfix: last coloured cluster sometimes wasn't shown
When no invalid clusters were present, the last cluster wouldn't
display in the results view. This occured because of a counting
error, because a loop always assumed a cluster containing invalid
points was always present.

The fix was to add an empty cluster (with id 0) to the map
containing cluster data, in the ClusterData constructor